pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendoor"
version = "4.0.0"
description = "OpenDoor AI coding assistant"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-genai",
    "pygments",
    "requests",
]

[project.scripts]
opendoor = "opendoor.core.app:main_entry"
